处理count(*)函数报错问题。

+++ src/main/java/com/example/api_project/mapper/PurchaseIntegrateMapper.java
@@ -20,9 +20,9 @@ public interface PurchaseIntegrateMapper {
     List<PurchaseIntegrate> queryAllByGoods(String goodsCode,Integer startRows, Integer pageSize);
     List<PurchaseIntegrate> queryAllBySupplier(String supplierCode,Integer startRows, Integer pageSize);
 
-    Long countByShop(PurchaseIntegrate purchaseIntegrate);
-    Long countByGoods(PurchaseIntegrate purchaseIntegrate);
-    Long countBySupplier(PurchaseIntegrate purchaseIntegrate);
+    Integer countByShop(PurchaseIntegrate purchaseIntegrate);
+    Integer countByGoods(PurchaseIntegrate purchaseIntegrate);
+    Integer countBySupplier(PurchaseIntegrate purchaseIntegrate);
 
 }
