+++ src/main/java/com/example/api_project/mapper/TransferIntegrateMapper.java
@@ -20,6 +20,6 @@ public interface TransferIntegrateMapper {
     List<TransferIntegrate> queryAllByShopOut(String shopCode, Integer startRows, Integer pageSize);
     List<TransferIntegrate> queryAllByGoods(String goodsCode,String InShopCode, String OutShopCode,Integer startRows, Integer pageSize);
 
-    Long countByGoods(TransferIntegrate transferIntegrate);
+    Integer countByGoods(TransferIntegrate transferIntegrate);
 }
